Fixed a pretty serious error
The system would crap out if the audio it was trying to load was already a mono signal. We don't want it to do that. We *definitely* don't want it to do that. Fixed.

Also, I kept selecting length(signal)-1 for some reason, so I missed the very last sample. I ain't doin that no more.

And I decided to keep some more of the variables instead of just clearing them all from memory. Good for debugging purposes. Or if the user wants to entertain themselves with pretty graphs.